chore(coordinator): Move admin.rs to routes module
